pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "my-langgraph-agent"
version = "0.1.0"
description = "A 2-node LangGraph agent running on a local Ollama model."
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0,<0.3.0",
    "langchain-ollama>=0.2.0",
    "langchain-core>=0.3.0,<0.4.0",
    "python-dotenv>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
